Calculate statistics on NURP tests
This calculates the standard deviation and z-score for each test NURP.

The z-score for a test NURP is it's working percentage minus the test's
overall mean divided by the stanard deviation. This tells us how much
better or worse a particular NURP is than the mean.

I would suggest filtering where standard deviation is > 0.5 or so, to
find NURPs that have some variation in their percentages, otherwise
you'll pick up tests that typically are 100% and went down to 99.8%
because of one failure out of a few dozen.