Correctly handle suite names for prow
Sippy prow is missing some synthetic test data because the suite isn't
removed like it is on TestGrid. I attempted to fix this in
https://github.com/openshift/sippy/pull/488, but I didn't realize
the synthetic test generation code was adding the sippy suite name to
the test name directly.

This ensures junits are correct: suite name only in the testsuite, and
where required for testgrid, joins them.