Sync pull request data from GitHub
This change links pull requests to job runs, allowing us to do
additional analysis. Pull requests are linked as both the PR being
tested, as well as the specific SHA -- meaning the analysis can be done
on commits that actually merged allow us to identify the most painful
and frequent failures.